+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added typing for the next config
+- added new parameter `enabled` to determ if google secrets should be loaded or not
 
 ## [0.0.1] - 2023-08-03